+++ nodejs/eks/nodegroup.ts
@@ -131,6 +131,11 @@ export interface NodeGroupBaseOptions {
      * public IPs.
      */
     nodeAssociatePublicIpAddress?: boolean;
+
+    /**
+     * Desired Kubernetes master / control plane version. If you do not specify a value, the latest available version is used.
+     */
+    version?: pulumi.Input<string>;
 }
 
 /**
@@ -310,17 +315,47 @@ ${customUserData}
 `;
         });
 
-    let amiId: pulumi.Input<string> = args.amiId!;
+    let amiId: any = args.amiId!;
+    const version: pulumi.Input<string> = args.version!;
     if (args.amiId === undefined) {
-        const eksWorkerAmi = aws.getAmi({
-            filters: [{
-                name: "name",
-                values: [ "amazon-eks-node-*" ],
-            }],
-            mostRecent: true,
+        const filters: { name: string; values: string[]}[] = [
+            {
+                name: "description",
+                values: [ "*linux*", "*Linux*" ],
+            },
+            {
+                name: "description",
+                values: [ "*k8s*/bin/linux/amd64*"],
+            },
+        ];
+
+        if (version !== undefined) {
+            filters.push(
+                {
+                    name: "name",
+                    values: [ "amazon-eks-node-" + version + "*" ],
+                },
+            );
+        } else {
+            filters.push(
+                {
+                    name: "name",
+                    values: [ "amazon-eks-node-*" ],
+                },
+            );
+        }
+
+        const eksWorkerAmiIds = aws.getAmiIds({
+            filters,
             owners: [ "602401143452" ], // Amazon
+            sortAscending: true,
         }, { parent: parent });
-        amiId = eksWorkerAmi.then(r => r.imageId);
+
+        const bestAmiId = eksWorkerAmiIds.then(r => r.ids.pop());
+        if (!bestAmiId) {
+            throw new Error("No Linux AMI Id was found.");
+        }
+        amiId = bestAmiId;
     }
 
     // Enable auto-assignment of public IP addresses on worker nodes for